requirements.txt
anytree
jsonpickle
pillow
PyQt5
PyOpenGL
numpy==1.22.3